migration/requirements.txt
pyjwt==2.4.0
psycopg==3.1.7